Move more functions to mixin
